added company dto

+++ src/modules/companies/repository/companies.repo.js
@@ -1,17 +1,20 @@
-const { QueryTypes,Sequelize } = require('sequelize');
+const { QueryTypes, Sequelize } = require('sequelize');
 import { RequestedFields } from '../../../RequestedFields';
 import { Company, CompanyMedia } from '../../../models';
 import Paginator from '../../../utils/Paginator';
 import PaginatedList from '../../../utils/PaginatedList';
+import CompanyDTO from '../../../dtos/CompanyDTO';
+import { CreateCompanyDTO, UpdateCompanyDTO } from '../../../dtos/CompanyMutationDTO'
+import { CreateMediaDTO } from '../../../dtos/MediaMutationDTO';
 
 const companiesRepository = db => {
     const requestedFields = new RequestedFields();
     const getFields = info => requestedFields.getFields(info, { keep: ["id", "user_id"], exclude: ["user", "medias"] })
-    const getFieldsWithSubfields = info => requestedFields.getFieldsWithSubfields(info, { keep: ["id", "user_id"], exclude: ["user","medias"] })
+    const getFieldsWithSubfields = info => requestedFields.getFieldsWithSubfields(info, { keep: ["id", "user_id"], exclude: ["user", "medias"] })
 
     const listRandomLogos = async (info, args) => {
         const fields = getFields(info)
-        const options = { attributes: fields,  order: [Sequelize.fn('RAND')]  }
+        const options = { attributes: fields, order: [Sequelize.fn('RAND')] }
         if (args.start != null && args.max != null) {
             options.offset = args.start;
             options.limit = args.max;
@@ -39,20 +42,21 @@ const companiesRepository = db => {
 
     const create = async args => {
         try {
-            const company = await Company.create(args.input, { include: { model: CompanyMedia, as: "medias" } })
+            const options = args.input.medias ? { include: { model: CompanyMedia, as: "medias" } } : {}
+            const company = await Company.create(new CreateCompanyDTO(args.input), options)
             await company.reload()
-            return company;    
+            return company;
         } catch (error) {
             console.error(error)
-            if (error.errors){
-                const errors = error.errors.map(e => e.message)                
+            if (error.errors) {
+                const errors = error.errors.map(e => e.message)
                 throw new Error(errors.toString())
-            }else {                
+            } else {
                 throw new Error(error.message)
             }
-            
+
         }
-        
+
     }
 
     const destroy = async args => {
@@ -69,42 +73,45 @@ const companiesRepository = db => {
 
         await db.sequelize.transaction(async transaction => {
             // chain all your queries here. make sure you return them.
-            const [companies, meta] = await Company.update(args.input, { where: { id: args.id }, returning: true }, { transaction })       
+            const [companies, meta] = await Company.update(new UpdateCompanyDTO(args.input), { where: { id: args.id }, returning: true, individualHooks: true }, { transaction })
 
-            await CompanyMedia.destroy({ where: { id: args.id } }, { transaction })
-    
-            for (const m of args.input.medias) {
-                await CompanyMedia.create({ id: args.id, media: m.media, url: m.url }, { transaction })
+            if (args.input.medias) {
+                for (const m of args.input.medias) {
+                    await CompanyMedia.destroy({ where: { id: args.id } }, { transaction })
+                    const mediaInput = { id: args.id, media: m.media, url: m.url }
+                    await CompanyMedia.create(new CreateMediaDTO(mediaInput), { transaction })
+                }
             }
-        })     
+        })
 
         const company = await Company.findOne({ where: { id: args.id }, include: { model: CompanyMedia, as: "medias" } })
 
         return company;
     }
 
-    const findAllPaginated = async (info, args) => {           
-        
-        const fields = getFieldsWithSubfields(info).get("rows")    
-        fields.push("user_id")                  
-        
+    const findAllPaginated = async (info, args) => {
+
+        const fields = getFieldsWithSubfields(info).get("companies")
+        fields.push("user_id")
+
         const totalRows = await Company.count()
 
         const paginator = new Paginator(args.limit, args.page, totalRows);
 
         const totalPages = paginator.getTotalPages();
 
-		const start = paginator.getStart();
+        const start = paginator.getStart();
 
-		const end = paginator.getEnd();		
+        const end = paginator.getEnd();
 
         const options = { attributes: fields, order: ['id'] }
         options.offset = start - 1;
         options.limit = args.limit;
 
-        const companies = await Company.findAll(options)
+        let companies = await Company.findAll(options)
+        companies = companies.map(c => new CompanyDTO(c))
 
-        const paginatedList = new PaginatedList(companies, start, end, totalPages, paginator.getCurrentPage(), paginator.getLimitRows(), paginator.getMaxRows())
+        const paginatedList = new PaginatedList('companies', companies, start, end, totalPages, paginator.getCurrentPage(), paginator.getLimitRows(), paginator.getMaxRows())
         return paginatedList;
     }
